// interfaces/entry.ts
export interface Entry {
  _id: string;
  description: string;
  status: EntryStatus;
  createdAt: number;
}

export type EntryStatus = "Pending" | "In-Progress" | "Finished";
// Generated by https://quicktype.io

// export interface EntryRes {
//   _id: string;
//   description: string;
//   status: string;
//   __v: number;
// }